fs: move ColumnHeader
